Code rearrangement
Merge Simulation_context and Simulation_context_base to simplify
the class hierarchy.
